+++ HISTORY.md
@@ -9,6 +9,7 @@
   public API.
 * Deprecate `Wavefront::User` class, as [the user API is now
   deprecated](https://docs.wavefront.com/2020.06.x_release_notes.html#obsolete-and-deprecated-apis)
+* Add `Wavefront::Role` class, for managing roles.
 
 ## 4.0.0 (2020-02-17)
 * Drop support for Ruby 2.3. (Breaking change.)